requirements.txt
playwright==1.49.0
python-dotenv==1.0.0
fake-useragent==1.4.0
playwright-stealth==1.0.6
